chore: rename socket folder to sockets
